// frontend/src/pages/Coping/models/CopingResult.ts
export interface CopingResult {
    date_time: string;
    confrontation: number;
    distancing: number;
    self_control: number;
    seeking_social_support: number;
    taking_responsibility: number;
    escape_avoidance: number;
    problem_solving_planning: number;
    positive_reassessment: number;
    confrontation_tpoint: number;
    distancing_tpoint: number;
    self_control_tpoint: number;
    seeking_social_support_tpoint: number;
    taking_responsibility_tpoint: number;
    escape_avoidance_tpoint: number;
    problem_solving_planning_tpoint: number;
    positive_reassessment_tpoint: number;
}